server/requirements_enhanced.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
mysql-connector-python==8.1.0
requests==2.31.0
geopy==2.3.0
joblib==1.3.2
python-dotenv==1.0.0
APScheduler==3.10.4
redis==4.6.0
celery==5.3.1